requirements.txt
librosa
transformers
numpy
torch
torchaudio
pympi-ling
pyctcdecode
